requirements.txt
Django==1.11.4
olefile==0.44
Pillow==4.2.1
pytz==2017.2
djangorestframework
